Upgrade to webpack v5, and use @fork-of/git-branch
https://github.com/webdoc-labs/webdoc/issues/73

Upgraded webpack to v5 in @webdoc/default-template's build. Upgraded css-loader, sass, sass-loader to work with webpack 5. Migrated away from optimize-css-assets-webpack-plugin and now use css-minimizer-webpack-plugin (former is not maintained). Installed webpack-cli and use it for the build command, and got rid Gulp from @webdoc/default-template.

Gulp causes an indirect dependency to urix, which is a deprecated; it will be completely removed in a future commit.

Moved to @fork-of/git-branch because git-branch was no longer actively maintained and the fork got rid of the urix dependency. It supposedly also added support for git worktree.